refactor: Refactor lstm method using factory mode
